locomotive now change direction along railway

+++ src/meowtro/metro_system/railway/RailwayRealizer.java
@@ -3,6 +3,6 @@ package meowtro.metro_system.railway;
 import meowtro.Position;
 
 public interface RailwayRealizer {
-    public int parsePositionToAbstractPosition(Position p); 
+    public double parsePositionToAbstractPosition(Position p); 
     public Position parseAbstractPositionToPosition(double abstractPosition); 
 }
